chore: pin Cordova plugins
Fork device-information so that a tag can be set. However, it's unmaintained and
ideally, a replacement should be found.

Remove unused local-notifications. This was included when we were researching
background service implementations, in our case for background stock count sync,
which is still an open issue (#10) and should be revisited using
eHealthAfrica/cordova-plugin-background-sync.

Closes #206.